Common: Rename abstract_event.py -> abstract_agent_event.py
