+++ monkey/common/events/__init__.py
@@ -1,2 +1,2 @@
-from .abstract_event import AbstractAgentEvent
+from .abstract_agent_event import AbstractAgentEvent
 from .credentials_stolen_events import CredentialsStolenEvent
